feature(chapter7): Implement the sell orders flow
Implement the flow described in Figure 2.4 for selling stocks.
There is still no persistence but the the docker-compose included allows
to spawn the services responsible to collaborate in the feature.

The gateway receives the http request and forwards it via rpc to the
orders service. Then a mix of emission of events and RPC calls allows
completing the flow.